feat (product variant): service to get product variant details with price promotion

+++ dto/product_variant_price_res.go
@@ -3,6 +3,6 @@ package dto
 type ProductVariantPriceRes struct {
 	ProductID        uint                          `json:"product_id" binding:"required"`
 	ProductName      string                        `json:"product_name" binding:"required"`
-	ProductPromotion uint                          `json:"product_promotion" binding:"required"`
+	ProductPromotion float64                       `json:"product_promotion" binding:"required"`
 	ProductVariant   []*ProductVariantPromotionRes `json:"product_variant" binding:"required"`
 }
